Removed unnecessary files, renamed Jobs files to be Action themed, fixed imports and navigation
