+++ CadetLinks/src/navigation/index.tsx
@@ -6,7 +6,7 @@ import { Image } from 'react-native';
 import { HomePage } from './screens/HomePage/Home';
 import { Login } from "./screens/LoginPage/Login";
 import { Profile } from './screens/ProfilePage/Profile';
-import { Jobs } from './screens/JobsPage/Jobs';
+import { Actions } from './screens/ActionsPage/Actions';
 import { Events } from './screens/EventsPage/EventScreen';
 import { NotFound } from './screens/NotFound';
 import { Search } from "./screens/SearchPage/Search";
@@ -54,10 +54,10 @@ const HomeTabs = createBottomTabNavigator({
         ...createTabIcon(calendar),
       },
     },
-    Jobs: {
-      screen: Jobs,
+    Actions: {
+      screen: Actions,
       options: {
-        title: "Jobs",
+        title: "Actions",
         headerShown: false,
         tabBarIcon: ({ color, size }) => (
           <Image
@@ -68,7 +68,6 @@ const HomeTabs = createBottomTabNavigator({
         ),
       },
     },
-    
     Profile: {
       screen: Profile,
       options: {
