+++ CadetLinks/src/navigation/screens/ActionsPage/Components/UploadDocsModal.tsx
@@ -11,7 +11,7 @@ import {
 import { Ionicons } from "@expo/vector-icons";
 import { generalStyles as g_styles, generalStyles } from "../../../../styles/GeneralStyles";
 import { DarkColors as colors } from "../../../../styles/colors";
-import { jobStyles as styles } from "../../../../styles/JobStyles";
+import { actionStyles as styles } from "../../../../styles/ActionStyles";
 import { eventsStyles } from "../../../../styles/EventStyles";
 import { UploadDocsModalProps } from "../../../../assets/types";
