+++ CadetLinks/src/navigation/screens/ActionsPage/Components/ViewDocumentModal.tsx
@@ -14,7 +14,7 @@ import {
 import { Ionicons } from "@expo/vector-icons";
 import { generalStyles as g_styles } from "../../../../styles/GeneralStyles";
 import { DarkColors as colors } from "../../../../styles/colors";
-import { jobStyles as styles } from "../../../../styles/JobStyles";
+import { actionStyles as styles } from "../../../../styles/ActionStyles";
 
 interface DocumentItem {
     dbKey: string;
